requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
click==8.3.0
colorama==0.4.6
Django==5.1.4
django-environ==0.12.0
django-ninja==1.4.3
h11==0.16.0
httptools==0.7.1
idna==3.10
pillow==11.3.0
psycopg2-binary==2.9.11
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
debugpy==1.8.17